迁移代码

+++ aiurt-devops/src/main/java/com/aiurt/boot/modules/manage/controller/SpecialSituationController.java
@@ -1,14 +1,13 @@
 package com.aiurt.boot.modules.manage.controller;
 
 import cn.hutool.core.util.ObjectUtil;
+import com.aiurt.common.aspect.annotation.AutoLog;
 import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.aiurt.boot.common.aspect.annotation.AutoLog;
-import com.aiurt.boot.common.system.query.QueryGenerator;
 import com.aiurt.boot.common.system.vo.LoginUser;
 import com.aiurt.boot.common.util.DateUtils;
 import com.aiurt.boot.common.util.oConvertUtils;
@@ -28,6 +27,8 @@ import io.swagger.annotations.ApiModelProperty;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shiro.SecurityUtils;
+import org.jeecg.common.api.vo.Result;
+import org.jeecg.common.system.query.QueryGenerator;
 import org.jeecgframework.poi.excel.ExcelImportUtil;
 import org.jeecgframework.poi.excel.def.NormalExcelConstants;
 import org.jeecgframework.poi.excel.entity.ExportParams;
@@ -441,7 +442,7 @@ public class SpecialSituationController {
 
     @GetMapping(value = "/getSpecialSituationNum")
     public Result getSpecialSituationNum(@RequestParam(name = "startTime", required = false) String startTime, @RequestParam(name = "endTime", required = false) String endTime) {
-        int count = 0;
+        long count = 0;
         if (oConvertUtils.isNotEmpty(startTime) && oConvertUtils.isNotEmpty(endTime)) {
             count = specialSituationService.count(new QueryWrapper<SpecialSituation>()
                     .between("publish_time", startTime, endTime)
